extract api type definitions to types module
