+++ src/frontend/src/components/disconnect-provider-button.tsx
@@ -1,4 +1,5 @@
-import { Connection, NotificationType, state } from '@/library/api';
+import { state } from '@/library/api';
+import { Connection, NotificationType } from '@/library/types';
 import React, { useState } from 'react';
 
 interface DisconnectProviderButtonComponent {
