+++ src/frontend/src/library/api.ts
@@ -1,8 +1,8 @@
-import { CarrierSettings, References, Shipment, Purplship, Address, Parcel, Customs } from '@purplship/purplship';
+import { ConnectionData, DefaultTemplates, LabelData, Log, PaginatedConnections, PaginatedLogs, PaginatedShipments, PaginatedTemplates, RequestError, Template, UserInfo, Notification, Collection } from '@/library/types';
+import { References, Shipment, Purplship, Address, Parcel, Customs } from '@purplship/purplship';
 import { useEffect, useState } from 'react';
-import { Subject, BehaviorSubject } from 'rxjs';
-import { distinct } from 'rxjs/operators';
-import { RequestError } from '@/library/types';
+import { Subject, BehaviorSubject, interval } from 'rxjs';
+import { distinct, throttle } from 'rxjs/operators';
 
 // Collect API token from the web page
 const INITIAL_TOKEN = collectToken();
@@ -23,79 +23,9 @@ const DEFAULT_PAGINATED_RESULT = {
     fetched: false,
 };
 
-export interface UserInfo {
-    full_name: string | null;
-    email: string | null;
-    readonly is_staff: boolean;
-}
-
-export interface Connection extends Omit<CarrierSettings, 'id' | 'carrier_name'> {
-    id: string | null | undefined;
-    carrier_name: CarrierSettings.CarrierNameEnum | 'none';
-    [property: string]: any;
-}
-
-export interface ConnectionData {
-    carrier_name: CarrierSettings.CarrierNameEnum;
-    carrier_config: Partial<Connection>;
-}
-
-export interface Log {
-    id: string;
-    requested_at: string;
-    response_ms: string;
-    path: string;
-    view: string;
-    view_method: string;
-    remote_addr: string;
-    host: string;
-    method: string;
-    query_params: string;
-    data: string;
-    response: string;
-    status_code: string;
-}
-
-export interface Template {
-    id?: string;
-    label?: string;
-    address?: Address;
-    customs?: Customs;
-    parcel?: Parcel;
-}
-
-interface PaginatedContent<T> {
-    count: Number;
-    url?: string | null;
-    next?: string | null;
-    previous?: string | null;
-    results: T[];
-    fetched?: boolean;
-}
-
-export interface PaginatedLogs extends PaginatedContent<Log> { }
-export interface PaginatedShipments extends PaginatedContent<Shipment> { }
-export interface PaginatedTemplates extends PaginatedContent<Template> { }
-export interface PaginatedConnections extends PaginatedContent<Connection> { }
-
-
-export enum NotificationType {
-    error = "is-danger",
-    warning = "is-warning",
-    info = "is-info",
-    success = "is-success"
-}
-
-export interface Notification {
-    type: NotificationType;
-    message: string | Error | RequestError;
-}
-
-export interface LabelData {
-    shipment: Shipment;
-}
-
 class AppState {
+    private Queue: Collection<Subject<any>> = {};
+
     private token$: BehaviorSubject<string> = new BehaviorSubject<string>(INITIAL_TOKEN);
     private user$: BehaviorSubject<UserInfo> = new BehaviorSubject<UserInfo>({} as UserInfo);
     private shipments$: Subject<PaginatedShipments> = new Subject<PaginatedShipments>();
