+++ src/frontend/src/views/app.tsx
@@ -15,8 +15,9 @@ import BoardFooter from '@/components/footer/board-footer';
 import Navbar from '@/components/navbar/navbar';
 import Notifier from '@/components/notifier';
 import LocationTitle from '@/components/location-title';
-import { PaginatedLogs, state, UserInfo, } from '@/library/api';
-import { Logs, Reference, User } from '@/library/context';
+import { state} from '@/library/api';
+import { DefaultTemplates, PaginatedLogs, PaginatedTemplates, UserInfo } from '@/library/types';
+import { Templates, Logs, Reference, User, ParcelTemplates, AddressTemplates } from '@/library/context';
 import 'prismjs';
 import 'prismjs/components/prism-json';
 import 'prismjs/themes/prism.css';
